Change (#118): Move data-set functions to another lib
Move all current data-set functions to the already
existing but unused libKitsunemimiHanamiFiles lib.
