+++ src/Hanami/src/api/http/v1/data_files/mnist/finalize_mnist_data_set.cpp
@@ -25,8 +25,9 @@
 #include <hanami_root.h>
 #include <database/data_set_table.h>
 #include <core/temp_file_handler.h>
-#include <core/data_set_files/data_set_file.h>
-#include <core/data_set_files/image_data_set_file.h>
+
+#include <libKitsunemimiHanamiFiles/data_set_files/data_set_file.h>
+#include <libKitsunemimiHanamiFiles/data_set_files/image_data_set_file.h>
 
 #include <libKitsunemimiCrypto/common.h>
 #include <libKitsunemimiJson/json_item.h>
