--- src/Hanami/src/core/data_set_files/table_data_set_file.h
@@ -1,49 +1,0 @@
-/**
- * @file        table_data_set_file.cpp
- *
- * @author      Tobias Anker <tobias.anker@kitsunemimi.moe>
- *
- * @copyright   Apache License Version 2.0
- *
- *      Copyright 2022 Tobias Anker
- *
- *      Licensed under the Apache License, Version 2.0 (the "License");
- *      you may not use this file except in compliance with the License.
- *      You may obtain a copy of the License at
- *
- *          http://www.apache.org/licenses/LICENSE-2.0
- *
- *      Unless required by applicable law or agreed to in writing, software
- *      distributed under the License is distributed on an "AS IS" BASIS,
- *      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *      See the License for the specific language governing permissions and
- *      limitations under the License.
- */
-
-#ifndef HANAMI_TABLEDATASETFILE_H
-#define HANAMI_TABLEDATASETFILE_H
-
-#include <core/data_set_files/data_set_file.h>
-
-class TableDataSetFile
-        : public DataSetFile
-{
-public:
-    TableDataSetFile(const std::string &filePath);
-    TableDataSetFile(Kitsunemimi::BinaryFile* file);
-    ~TableDataSetFile();
-    bool updateHeader();
-    float* getPayload(uint64_t &payloadSize,
-                      const std::string &columnName = "");
-
-    void print();
-
-    TableTypeHeader tableHeader;
-    std::vector<TableHeaderEntry> tableColumns;
-
-protected:
-    void initHeader();
-    void readHeader(const uint8_t* u8buffer);
-};
-
-#endif // HANAMI_TABLEDATASETFILE_H
