+++ src/libraries/libKitsunemimiHanamiFiles/src/src.pro
@@ -10,23 +10,23 @@ LIBS += -L../../libKitsunemimiCommon/src/debug -lKitsunemimiCommon
 LIBS += -L../../libKitsunemimiCommon/src/release -lKitsunemimiCommon
 INCLUDEPATH += ../../libKitsunemimiCommon/include
 
-LIBS += -L../../libKitsunemimiHanamiCommon/src -lKitsunemimiHanamiCommon
-LIBS += -L../../libKitsunemimiHanamiCommon/src/debug -lKitsunemimiHanamiCommon
-LIBS += -L../../libKitsunemimiHanamiCommon/src/release -lKitsunemimiHanamiCommon
-INCLUDEPATH += ../../libKitsunemimiHanamiCommon/include
+LIBS += -L../../libKitsunemimiJson/src -lKitsunemimiJson
+LIBS += -L../../libKitsunemimiJson/src/debug -lKitsunemimiJson
+LIBS += -L../../libKitsunemimiJson/src/release -lKitsunemimiJson
+INCLUDEPATH += ../../libKitsunemimiJson/include
 
 INCLUDEPATH += $$PWD \
                $$PWD/../include
 
-HEADERS += \
-    ../include/libKitsunemimiHanamiFiles/data_set_files/data_set_file.h \
-    ../include/libKitsunemimiHanamiFiles/data_set_files/image_data_set_file.h \
-    ../include/libKitsunemimiHanamiFiles/data_set_files/table_data_set_file.h \
-    ../include/libKitsunemimiHanamiFiles/structs.h
-
 SOURCES += \
     data_set_files/data_set_file.cpp \
+    data_set_files/data_set_functions.cpp \
     data_set_files/image_data_set_file.cpp \
     data_set_files/table_data_set_file.cpp
 
+HEADERS += \
+    ../include/libKitsunemimiHanamiFiles/data_set_files/data_set_file.h \
+    ../include/libKitsunemimiHanamiFiles/data_set_files/data_set_functions.h \
+    ../include/libKitsunemimiHanamiFiles/data_set_files/image_data_set_file.h \
+    ../include/libKitsunemimiHanamiFiles/data_set_files/table_data_set_file.h
